DBZ-2975: Move Partition to io.debezium.pipeline.spi
