+++ debezium-core/src/main/java/io/debezium/pipeline/spi/Partition.java
@@ -3,7 +3,7 @@
  *
  * Licensed under the Apache Software License version 2.0, available at http://www.apache.org/licenses/LICENSE-2.0
  */
-package io.debezium.connector.common;
+package io.debezium.pipeline.spi;
 
 import java.util.Map;
 import java.util.Set;
